requirements.txt
python-dotenv
requests
streamlit
pandas
plotly
passlib[bcrypt]

